BE/PBL6/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==5.1
django-anymail==11.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-crispy-forms==2.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
